fix(lint): Remove hard coded reference to /sdcard
The lint tool mark this is error so fix it, and replace hard code '/'
of path with File tool.

Verify by new installation without rime folder, and test almost all
functional, it works.

Though it's not perfect but it works, some verify and polish need to
be done before merge.